+++ src/main/java/in/sakthi/util/ConnectionUtil.java
@@ -3,27 +3,27 @@ package in.sakthi.util;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
+import java.sql.Statement;
 
 public class ConnectionUtil {
-	  /**
-	   * Database Connection
-	   */
-	 private ConnectionUtil() {
-		    throw new IllegalStateException("ConnectionUtil");
-		  }
+	/**
+	 * Database Connection
+	 */
+	private ConnectionUtil() {
+		throw new IllegalStateException("ConnectionUtil");
+	}
 
-	    private static final String DRIVER_CLASS_NAME = System.getenv("spring.datasource.driver-class-name");
-		private static final String DATABASE_NAME = System.getenv("spring.datasource.url");
-		private static final String USERNAME = System.getenv("spring.datasource.username");
-		private static final String PASSWORD =System.getenv( "spring.datasource.password");
-		private static final String DB_URL= DATABASE_NAME;
+	private static final String DRIVER_CLASS_NAME = System.getenv("spring.datasource.driver-class-name");
+	private static final String DATABASE_NAME = System.getenv("spring.datasource.url");
+	private static final String USERNAME = System.getenv("spring.datasource.username");
+	private static final String PASSWORD = System.getenv("spring.datasource.password");
+	private static final String DB_URL = DATABASE_NAME;
 
-	public static Connection getConnection()
-	{
-	Connection connection = null;	
+	public static Connection getConnection() {
+		Connection connection = null;
 		try {
 			Class.forName(DRIVER_CLASS_NAME);
-			connection = DriverManager.getConnection(DB_URL,USERNAME , PASSWORD);
+			connection = DriverManager.getConnection(DB_URL, USERNAME, PASSWORD);
 			System.out.println(connection);
 		} catch (ClassNotFoundException | SQLException e) {
 			e.printStackTrace();
@@ -32,8 +32,17 @@ public class ConnectionUtil {
 		return connection;
 	}
 
-	
-		
+	public static void close(Statement pst, Connection Connection) throws SQLException {
+		try {
+			if (pst != null) {
+				pst.close();
+			}
+			if (Connection != null) {
+				Connection.close();
+			}
+		} catch (SQLException e) {
+			e.printStackTrace();
+		}
 	}
 
-
+}
