+++ src/test/java/in/sakthi/TestChannelService.java
@@ -9,10 +9,10 @@ import in.sakthi.service.ChannelService;
 public class TestChannelService {
 
 	public static void main(String[] args) throws SQLException {
-		
+
 		List<Channel> channel = ChannelService.getList();
 		System.out.println();
-      
+
 	}
 
 }
